greatly simplify and optimize implementation
stop inducing embeddings, we don't need to ― this is an extractive
approach, and embeddings largely capture the same information for our
candidate set, so using them is superfluous; they are only useful for
their intended application, i.e., to emplace arbitrary sequences of
embedded unigrams within the same semantic high-dimensional vector
space. All of our candidates are permutations that appear in the
original text, not arbitrary prose, so we don't have to bother with
that.